feat(Header): Refactor header component and extract account menu functionality
- Removed unused imports and code related to dark mode toggle
- Extracted account menu functionality into a separate component called AccountMenu
- Updated Header component to use the new AccountMenu component for displaying account settings
