+++ src/Layout/Header.jsx
@@ -1,40 +1,10 @@
-import { useAuth0 } from "@auth0/auth0-react";
-import {
-  Toolbar,
-  Box,
-  IconButton,
-  Typography,
-  Avatar,
-  Tooltip,
-  Menu,
-  MenuItem,
-} from "@mui/material";
+import { Toolbar, Box, IconButton, Typography } from "@mui/material";
 import AppBar from "@mui/material/AppBar";
 import MenuIcon from "@mui/icons-material/Menu";
-import Button from "@mui/material/Button";
-import { useContext, useState } from "react";
-import DarkModeContext from "../contexts/AppContext";
-import BedtimeOutlinedIcon from "@mui/icons-material/BedtimeOutlined";
-import BedtimeOffOutlinedIcon from "@mui/icons-material/BedtimeOffOutlined";
+import { useState } from "react";
+import AccountMenu from "../components/AccountMenu";
 
 const Header = () => {
-  const { isDarkMode, toggleDarkMode } = useContext(DarkModeContext);
-  const { isAuthenticated, loginWithRedirect, logout, user } = useAuth0();
-
-  const getInitials = () => {
-    return user?.name
-      .split(" ")
-      .map((n) => n[0])
-      .join("")
-      .toUpperCase();
-  };
-
-  const [anchorEl, setAnchorEl] = useState(null);
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
   return (
     <AppBar position="static" elevation={2}>
       <Toolbar>
@@ -51,84 +21,7 @@ const Header = () => {
           ToDo
         </Typography>
         <Box sx={{ flexGrow: 1 }} />
-        <Tooltip title="Account settings" placement="left">
-          <IconButton
-            onClick={(e) => setAnchorEl(e.currentTarget)}
-            size="small"
-          >
-            <Avatar src={user?.picture || null}>{getInitials()}</Avatar>
-          </IconButton>
-        </Tooltip>
-        <Menu
-          anchorEl={anchorEl}
-          open={Boolean(anchorEl)}
-          onClose={handleClose}
-          onClick={handleClose}
-          PaperProps={{
-            elevation: 0,
-            sx: {
-              overflow: "visible",
-              filter: "drop-shadow(0px 2px 8px rgba(0,0,0,0.32))",
-              mt: 1.5,
-              "& .MuiAvatar-root": {
-                width: 32,
-                height: 32,
-                ml: -0.5,
-                mr: 1,
-              },
-              "&:before": {
-                content: '""',
-                display: "block",
-                position: "absolute",
-                top: 0,
-                right: 14,
-                width: 10,
-                height: 10,
-                bgcolor: "background.paper",
-                transform: "translateY(-50%) rotate(45deg)",
-                zIndex: 0,
-              },
-            },
-          }}
-          transformOrigin={{ horizontal: "right", vertical: "top" }}
-          anchorOrigin={{ horizontal: "right", vertical: "bottom" }}
-        >
-          <MenuItem>
-            <Typography sx={{ p: "8px", fontSize: 14, fontWeight: "500" }}>
-              Profile
-            </Typography>
-          </MenuItem>
-          <MenuItem>
-            {" "}
-            <IconButton
-              onClick={() => toggleDarkMode()}
-              color="inherit"
-              aria-label="Toggle Dark Mode"
-            >
-              {isDarkMode ? (
-                <BedtimeOffOutlinedIcon />
-              ) : (
-                <BedtimeOutlinedIcon />
-              )}
-            </IconButton>
-          </MenuItem>
-          <MenuItem>
-            {isAuthenticated ? (
-              <Button
-                color="inherit"
-                onClick={() =>
-                  logout({ logoutParams: { returnTo: window.location.origin } })
-                }
-              >
-                Log Out
-              </Button>
-            ) : (
-              <Button color="inherit" onClick={() => loginWithRedirect()}>
-                Login
-              </Button>
-            )}
-          </MenuItem>
-        </Menu>
+        <AccountMenu />
       </Toolbar>
     </AppBar>
   );
